moved components into components folder, set up some routing
