--- frontend/src/UserDashboard.vue
@@ -1,12 +1,0 @@
-<template>
-    <div>
-      <h1>This will be the User Dashboard</h1>
-    </div>
-</template>
-
-<script>
-export default {
-    
-}
-</script>
-
