+++ frontend/src/components/FarmDashboard.vue
@@ -34,8 +34,8 @@
 </template>
 
 <script>
-import FeedBox from './components/FeedBox.vue';
-import SummaryBox from './components/SummaryBox.vue';
+import FeedBox from '../components/FeedBox.vue';
+import SummaryBox from '../components/SummaryBox.vue';
 
 export default {
   name: 'app',
