+++ frontend/src/components/UserDashboard.vue
@@ -1,15 +1,12 @@
 <template>
-  <div class="user-dashboard">
-    User dashboard here
-  </div>
+    <div>
+      <h1>This will be the User Dashboard</h1>
+    </div>
 </template>
 
 <script>
 export default {
-  name: 'UserDashboard'
+    
 }
 </script>
 
-<style>
-
-</style>
